fix app suspense 제거 (#309)
