+++ src/app/not-found.tsx
@@ -1,7 +1,7 @@
 export default function NotFound() {
   return (
     <div>
-      <p className=" text-center text-5xl p-5">Not Found</p>
+      <p className=" text-center text-5xl p-5 mt-40">Not Found</p>
     </div>
   );
 }
